test: Make TestMultiMachine.testDirectLogin compatible with beiboot/ssh
`cockpit-ssh` always considers both the global and the user known_hosts
files, and rejects a host if a non-matching key is present in either.
But with `ssh` the key in the user known_hosts file completely overrides
the global one.

In our case, the user known_hosts file is coming from the browser's
localStorage. So reset that before writing a bad key to the global file,
which is compatible with both backends.