src: move PolKit policy from bridge/ to systemd/
This isn't really a systemd file, but it fits in well with the rest of
the systems integration stuff in that directory.  Besides: this isn't
tied to the C bridge anymore: we use it from Python as well, and will be
getting rid of the C bridge soon.

There's a theoretical functional change here which makes no difference
in practice: we now install the policy file unconditionally.  We'll soon
remove the `--disable-polkit` option anyway, and the only two things
that call `--disable-polkit` don't call `make install`
(`tools/make-dist` and the flatpak build).